Add new ignition compatibility mode to stalks (#105)
New compatiblity mode added to multifunction stalks that allows users to transform rear wiper controls into ignition and starter switch. The on/off position is now tracked and button inputs are only sent when it actually changes + the buttons are now combined into just one.

The button doesn't fire again after toggling the momentary switch, normally used for rear windshield washer, so the whole thing can be used the same as a real ignition. Tested with Truck Simlators and ACC.